Implement FlatArray static
