+++ src/collections/Collections/Immutable.FlatArray.T/FlatArray.Factory.Implicit.cs
