+++ src/collections/Collections/Immutable.FlatArray.T/FlatArray.InnerArrayHelper.cs
