+++ src/collections/Collections/Immutable.FlatArray.T/FlatArray.To.Implicit.cs
